Code-Share `dra` 4AEA4.c
