+++ config/symbols.pspeu.dra.txt
@@ -557,6 +557,7 @@ g_ButtonMask = 0x0914C2C8;
 D_800B0884 = 0x09160B40;
 D_800B08A8 = 0x09160B68;
 g_LevelHPIncrease = 0x09160B90;
+g_PalEquipIcon = 0x091634C0;
 D_800CFE48 = 0x0916F738;
 D_800CF324 = 0x0916F800;
 D_800C52F8 = 0x09170960;
