+++ config/symbols.pspeu.txt
@@ -3,7 +3,7 @@ labs = 0x08909194;
 memcpy = 0x08909F84;
 srand = 0x0890B91C;
 rand = 0x0890B954;
-psp_sprintf = 0x0890C6FC;
+sprintf = 0x0890C6FC;
 strcat = 0x0890C928;
 STRCPY = 0x0890CA48;
 PadRead_PSP = 0x8919BE0;
